services/user-management/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic[email]==2.5.3
asyncpg==0.29.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
